Add protobuf benchmarks for varint encoding.
We add two sets of benchmarks:
1. This benchmark uses increasingly large Any messages, which are common for storage systems.
For database like services, which are agnostic to the type being stored, Proto2 extensions could be used.
However, Proto3 drops unknown fields and most of extensions, so Any must be used.
Any types are also used as the arbitrary payload mechanism for RPC metadata.
2. Another benchmark uses the barest possible proto, meant to show the overhead of the encoding system.
This is the default cost of encoding a proto with absolutely no fields.

PiperOriGin-RevId: 135044506