proto: Add a descriptor subpackage.
This provides a more reasonable API for obtaining a FileDescriptorProto and
DescriptorProto for a given proto.Message — a process that is currently possible
(but undocumented) using the public proto API.

The major use case for obtaining a DescriptorProto is to decode custom message
options, which are encoded as extensions on the MessageOptions submessage.
(See https://developers.google.com/protocol-buffers/docs/proto#customoptions.)

Fixes https://github.com/golang/protobuf/issues/179.
PiperOriGin-RevId: 139356528