[dev] protoc-gen-go: reorganize, fix testdata directory
"make test" now compiles all the source files in testdata/ using the
protoc-gen-go in the current working directory and compares the output
to golden versions, "make regenerate" rebuilds the golden files.

Add a go_package option to each proto source file. Put the sources for
each package in the proper directory.

Add a golden_test.go which arranges to compile each source file using
the compiler in the working tree.

This does not touch the content of any of the sources in testdata/
other than to add go_package options and fix up import paths.

Change-Id: Iea5bef9bba626116b8ce5ea136a15f3cff4f5bcc