protoc-gen-go: fix symbol names in comments on public imports
Some aliases generated for public imports include comments indicating
the source of the import. Fix these comments to match the name of the
imported symbol. Example changes:

    -// Submessage from public import import_public/sub/a.proto
    +// M_Submessage from public import import_public/sub/a.proto
     type M_Submessage = sub.M_Submessage

    -// Subenum from public import import_public/sub/a.proto
    +// M_Subenum from public import import_public/sub/a.proto
     type M_Subenum = sub.M_Subenum

    -// Submessage_Subenum from public import import_public/sub/a.proto
    +// M_Submessage_Submessage_Subenum from public import import_public/sub/a.proto
     type M_Submessage_Submessage_Subenum = sub.M_Submessage_Submessage_Subenum

Drop the "ignoring public import of {symbol} from {file}..." comment.
This comment can be somewhat confusing, since it doesn't explain why the
import is being ignored, and it's not easy in the current generator to
get the symbol name right.